+++ src/components/passwordfieldinput/index.tsx
@@ -1,52 +1,21 @@
 import React from 'react'
+import { filledTheme, outlinedTheme, standardTheme } from '../themes/Index'
 // import passwordshoweye from "./Image/passwordshoweye.svg"
 // import passwordhideeye from "./Image/passwordhideeye.svg"
 const PasswordField = (props: any) => {
 
-  const outlinedTheme = {
-    minHeight: '50px',
-    minWidth: '200px',
+  const errorTheme = {
     border: props.error ? '1px solid #ff0505' : '1px solid #000000',
-    outline: 'none'
   }
 
-  const standardTheme = {
-    minHeight: '50px',
-    minWidth: '200px',
-    boxSizing: 'border-box',
-    borderBottom: props.error ? '1px solid #ff0505' : '1px solid #000000',
-    borderTop: 'none',
-    borderLeft: 'none',
-    borderRight: 'none',
-    outline: 'none',
-    background: 'transparent',
-    WebkitTransition: '0.5s',
-    transition: '0.5s'
-  }
-
-  const filledTheme = {
-    minHeight: '50px',
-    minWidth: '200px',
-    boxSizing: 'border-box',
-    borderBottom: props.error ? '1px solid #ff0505' : '1px solid #000000',
-    borderTop: 'none',
-    borderLeft: 'none',
-    borderRight: 'none',
-    outline: 'none',
-    background: '#d7d7d74f',
-    WebkitTransition: '0.5s',
-    transition: '0.5s',
-    borderTopLeftRadius: '4px',
-    borderTopRightRadius: '4px'
-  }
   return (
     <>
       <p className={props.lableClass}>{props.Label}</p>
       <div style={{ position: 'relative', width: 'fit-content' }}>
         <input
-          style={props.variant === 'standard' ? standardTheme
-            : props.variant === 'filled' ? filledTheme
-              : outlinedTheme}
+          style={props.variant === 'standard' ? { ...standardTheme, ...errorTheme }
+            : props.variant === 'filled' ? { ...filledTheme, ...errorTheme }
+              : { ...outlinedTheme, ...errorTheme }}
           id={props.id}
           value={props.value}
           name={props.name}
